Add a first cut of a Dockerfile
We had some networking challenges at the most recent workshop at the
stage where we run `bundle install`. I'm playing with an idea where we
prepare an image ahead of time which has Rails and all of its
dependencies installed ahead of time, and we use that image as our base.
That way, the `bundle install` step will be super fast since all of the
gems are already available in the image.

Planning to set up the docker image to auto-build whenever we push to
this repository.

https://hub.docker.com/r/railsbridgenyc/rails/